Extract renderLayoutPage to templates/layout.dart
