+++ app/lib/frontend/templates.dart
@@ -24,6 +24,7 @@ import 'models.dart';
 import 'static_files.dart';
 import 'template_consts.dart';
 import 'templates/_cache.dart';
+import 'templates/layout.dart';
 
 String _escapeAngleBrackets(String msg) =>
     const HtmlEscape(HtmlEscapeMode.element).convert(msg);
@@ -751,68 +752,6 @@ class TemplateService {
     return templateCache.renderTemplate('mini_list', values);
   }
 
-  /// Renders the `views/layout.mustache` template.
-  String renderLayoutPage(
-    PageType type,
-    String contentHtml, {
-    @required String title,
-    String pageDescription,
-    String faviconUrl,
-    String canonicalUrl,
-    String platform,
-    SearchQuery searchQuery,
-    bool includeSurvey = true,
-    bool noIndex = false,
-  }) {
-    final queryText = searchQuery?.query;
-    final String escapedSearchQuery =
-        queryText == null ? null : htmlEscape.convert(queryText);
-    String platformTabs;
-    if (type == PageType.landing) {
-      platformTabs = renderPlatformTabs(platform: platform, isLanding: true);
-    } else if (type == PageType.listing) {
-      platformTabs =
-          renderPlatformTabs(platform: platform, searchQuery: searchQuery);
-    }
-    final searchSort = searchQuery?.order == null
-        ? null
-        : serializeSearchOrder(searchQuery.order);
-    final platformDict = getPlatformDict(platform);
-    final isRoot = type == PageType.landing && platform == null;
-    final values = {
-      'no_index': noIndex,
-      'static_assets_dir': staticUrls.staticPath,
-      'static_assets': staticUrls.assets,
-      'favicon': faviconUrl ?? staticUrls.smallDartFavicon,
-      'canonicalUrl': canonicalUrl,
-      'pageDescription': pageDescription == null
-          ? defaultPageDescriptionEscaped
-          : htmlEscape.convert(pageDescription),
-      'title': htmlEscape.convert(title),
-      'search_platform': platform,
-      'search_query': escapedSearchQuery,
-      'search_query_placeholder': platformDict.searchPlatformPackagesLabel,
-      'search_sort_param': searchSort,
-      'platform_tabs_html': platformTabs,
-      'api_search_enabled': searchQuery?.isApiEnabled ?? true,
-      'landing_blurb_html': platformDict.landingBlurb,
-      // This is not escaped as it is already escaped by the caller.
-      'content_html': contentHtml,
-      'include_survey': includeSurvey,
-      'include_highlight': type == PageType.package,
-      'landing_banner': type == PageType.landing,
-      'landing_banner_image':
-          platform == 'flutter' ? 'flutter-packages.png' : 'dart-packages.png',
-      'landing_banner_alt':
-          platform == 'flutter' ? 'Flutter packages' : 'Dart packages',
-      'listing_banner': type == PageType.listing,
-      'package_banner': type == PageType.package,
-      'schema_org_searchaction_json':
-          isRoot ? json.encode(_schemaOrgSearchAction) : null,
-    };
-    return templateCache.renderTemplate('layout', values, escapeValues: false);
-  }
-
   /// Renders the `views/pagination.mustache` template.
   String renderPagination(PageLinks pageLinks) {
     final values = {
@@ -874,44 +813,6 @@ class TemplateService {
       });
     }
     return templateCache.renderTemplate('pkg/tags', {'tags': tags});
-  }
-
-  String renderPlatformTabs({
-    String platform,
-    SearchQuery searchQuery,
-    bool isLanding = false,
-  }) {
-    final String currentPlatform = platform ?? searchQuery?.platform;
-    Map platformTabData(String tabText, String tabPlatform) {
-      String url;
-      if (searchQuery != null) {
-        final newQuery = searchQuery.change(
-            platform: tabPlatform == null ? '' : tabPlatform);
-        url = newQuery.toSearchLink();
-      } else {
-        final List<String> pathParts = [''];
-        if (tabPlatform != null) pathParts.add(tabPlatform);
-        if (!isLanding) pathParts.add('packages');
-        url = pathParts.join('/');
-        if (url.isEmpty) {
-          url = '/';
-        }
-      }
-      return {
-        'text': tabText,
-        'href': _attrEscaper.convert(url),
-        'active': tabPlatform == currentPlatform
-      };
-    }
-
-    final values = {
-      'tabs': [
-        platformTabData('Flutter', KnownPlatforms.flutter),
-        platformTabData('Web', KnownPlatforms.web),
-        platformTabData('All', null),
-      ]
-    };
-    return templateCache.renderTemplate('platform_tabs', values);
   }
 }
 
@@ -1052,23 +953,6 @@ class PageLinks {
     }
   }
 }
-
-enum PageType {
-  landing,
-  listing,
-  package,
-}
-
-const _schemaOrgSearchAction = const {
-  '@context': 'http://schema.org',
-  '@type': 'WebSite',
-  'url': '${urls.siteRoot}/',
-  'potentialAction': const {
-    '@type': 'SearchAction',
-    'target': '${urls.siteRoot}/packages?q={search_term_string}',
-    'query-input': 'required name=search_term_string',
-  },
-};
 
 Map _schemaOrgPkgMeta(Package p, PackageVersion pv, AnalysisView analysis) {
   final Map map = {
